Initial commit
This program accepts a spreadsheet of historical financial data.  First column must be indexable and first row must be column headers.  After determining which data you would like to use, edit the entries in "indep_variables" and "dep_variables" lists to reflect the column headers.  The result of this program is the maximized Ordinary Least Squares fit model.  Future development will include printing the shift offsets for the maximized model.  These correspond to the day offsets that provide the maximum predictive power on the inputs.  

>> WARNING: This code takes a *very* long time to run.  It also does not guarantee any financial outcomes for the end-user <<